Add files via upload
This is a js file that creates a dynamic list of search suggestions (10 is the max) and wraps each them in an html <li> tag. The <li> tags are surrounded by an html  <ul> tag. As the user types in the search term,  possible matches come back with every "key up" event and are placed underneath the input field where the user can scroll through them and make a selection. If a selection is made the form is submitted automatically.